Added useReducer for state update
Added useReducer for state update
